+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # next
 * On GHC 7.11 and up, `Generics.Deriving.TH` uses the new type literal-based
   machinery
+* Backported `Generic(1)` instances added in GHC 7.11. Specifically, `Generic`
+  instances for `Complex` (`base-4.4` and later) `ExitCode`, and `Version`; and
+  `Generic1` instances for `Complex` (`base-4.4` and later) and `Proxy`
+  (`base-4.7` and later). Added `GEnum`, `GEq`, `GFoldable`, `GFunctor`, `GIx`,
+  `GShow`, and `GTraversable` instances for these datatypes where appropriate.
 
 # 1.9.0
 * Allow deriving of Generic1 using Template Haskell
@@ -8,7 +13,8 @@
 * Allow deriving of Generic(1) for constructor-less plain datatypes (but not
   data families, due to technical restrictions)
 * Support for unboxed representation types on GHC 7.11+
-* More `GCopoint`, `GEnum`, `GEq`, `GFoldable`, `GFunctor`, `GIx`, `GMonoid`, `GShow`, and `GTraversable` instances
+* More `GCopoint`, `GEnum`, `GEq`, `GFoldable`, `GFunctor`, `GIx`, `GMonoid`,
+  `GShow`, and `GTraversable` instances
 * The field accessors for the `(:+:)` type in `Generics.Deriving.Base` have
   been removed to be consistent with `GHC.Generics`
 * Ensure that TH generates definitions for isNewtype and packageName, if a